// src/shared/enums/error.enum.ts
export enum ErrorCode {
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  UNAUTHORIZED = 'UNAUTHORIZED',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  DUPLICATE_RESOURCE = 'DUPLICATE_RESOURCE',
  DUPLICATE_KEY_ERROR = 'DUPLICATE_KEY_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
}